+++ backend/core/src/auth/services/authz.service.ts
@@ -48,14 +48,20 @@ export class AuthzService {
       //  A User becomes a leasing agent for a given listing if he has a relation (M:N) with it.
       //  User side this is expressed by 'leasingAgentInListings' property.
       await Promise.all(
-        user?.leasingAgentInListings.map((listing: Listing) =>
-          e.addPermissionForUser(
+        user?.leasingAgentInListings.map((listing: Listing) => {
+          void e.addPermissionForUser(
             user.id,
             "application",
             `!r.obj || r.obj.listing_id == '${listing.id}'`,
             `(${authzActions.read}|${authzActions.create}|${authzActions.update}|${authzActions.delete})`
           )
-        )
+          void e.addPermissionForUser(
+            user.id,
+            "listing",
+            `!r.obj || r.obj.listing_id == '${listing.id}'`,
+            `(${authzActions.read}|${authzActions.update})`
+          )
+        })
       )
     }