+++ backend/core/src/listings/listings.service.ts
@@ -14,7 +14,6 @@ import { Pagination } from "nestjs-typeorm-paginate"
 import { Repository } from "typeorm"
 import { plainToClass } from "class-transformer"
 import { PropertyCreateDto, PropertyUpdateDto } from "../property/dto/property.dto"
-import { arrayIndex } from "../libs/arrayLib"
 import { addFilters } from "../shared/filter"
 
 @Injectable()
@@ -101,21 +100,6 @@ export class ListingsService {
       itemsPerPage: itemsPerPage,
       totalItems: totalItems,
       totalPages: Math.ceil(totalItems / itemsPerPage), // will be 1 if no pagination
-    }
-
-    // Get the application counts and map them to listings
-    if (origin === process.env.PARTNERS_BASE_URL) {
-      const counts = await this.listingRepository
-        .createQueryBuilder("listing")
-        .select("listing.id")
-        .loadRelationCountAndMap("listing.applicationCount", "listing.applications", "applications")
-        .getMany()
-
-      const countIndex = arrayIndex<Listing>(counts, "id")
-
-      listings.forEach((listing: Listing) => {
-        listing.applicationCount = countIndex[listing.id].applicationCount || 0
-      })
     }
 
     // TODO(https://github.com/CityOfDetroit/bloom/issues/135): Decide whether to remove jsonpath