+++ backend/core/types/src/backend-swagger.ts
@@ -4222,9 +4222,6 @@ export interface Listing {
   applicationConfig?: object
 
   /**  */
-  applicationCount?: number
-
-  /**  */
   displayWaitlistSize: boolean
 
   /**  */